# Remove cuda stream in awq gemm achieved 2.5x speed up