Removing cuda stream in awq both gemm and dequantize, somehow achieved max 2.5x speed up

Command for testing:
`python benchmarks/benchmark_throughput.py --input-len 1024 --output-len 64 --model TheBloke/Mistral-7B-Instruct-v0.2-AWQ --quantization awq --num-prompts 100 --max-model-len 8192 --dtype half`

| input len | output len | main reqs/s | main tokens/s | PR reqs/s | PR tokens/s | Speed Up |
|-|-|-|-|-|-|-|
| 1024 | 64 | 8.61 | 9365.52 | 10.19 | 11085.39 | 1.18 |
| 512 | 64 | 14.24 | 8202.31 | 18.94 | 10908.37 | 1.33 |
| 256 | 64 | 20.78 | 6648.15 | 32.55 | 10416.5 | 1.57 |
| 128 | 128 | 15.75 | 4031.75 | 33.53 | 8583.89 | 2.13 |
| 64 | 64 | 32.44 | 4152.96 | 72.48 | 9277.53 | 2.23 |
| 32 | 64 | 35.44 | 3402.48 | 89.3 | 8772.66 | 2.52 |
| 32 | 32 | 65 | 4160.22 | 65.84 | 42123.63 | 1.01 |