@Yard1 2024-02-20
This most likely will result in incorrect outputs. I would imagine the speedup comes from the fact that the cuda graphed operations are now ran in parallel with AWQ kernels, but that's definitely not correct. The stream simply has to be set there.