@zcnrex 2024-02-20
@Yard1 Thanks for your insight and sorry about the noise! Closed the PR. Could you share if there's a way to test correctness on vllm? Also could you share good resources to better understand cuda stream?