@Yard1 2024-02-20
The easiest way is to compare outputs of models with temperature 0 before and after changes - they should be identical. We also have tests that compare the outputs to reference (HF Transformers).

This is a good resource regarding streams - https://leimao.github.io/blog/CUDA-Stream/